Rename `ControlItem` -> `ChannelListItem` and associated classes
